+++ src/main/java/com/codefarm/farmer/controller/admin/AdminNoticeController.java
@@ -1,4 +1,4 @@
-package com.codefarm.farmer.controller;
+package com.codefarm.farmer.controller.admin;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
