+++ src/main/java/com/codefarm/farmer/controller/admin/AdminUserController.java
@@ -1,14 +1,14 @@
-package com.codefarm.farmer.controller;
+package com.codefarm.farmer.controller.admin;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/adminMain/*")
-public class AdminMainController {
+@RequestMapping("/adminUser/*")
+public class AdminUserController {
 
-    @GetMapping("/main")
-    public void adminMain() {
+    @GetMapping("/user")
+    public void adminUser() {
     }
 }
