+++ src/main/resources/static/css/admin/adminCommon/contents.css
