+++ src/main/resources/static/css/admin/adminCommon/page.css
