+++ src/main/resources/static/css/admin/adminNotice/notice.css
