+++ src/main/resources/static/js/admin/adminCommon/chartAPI.js
