+++ src/main/resources/static/js/admin/adminCommon/check.js
