+++ src/main/resources/static/js/admin/adminCommon/side_menu.js
