+++ src/main/resources/static/js/admin/adminCommon/summerNote.js
