+++ src/main/resources/static/js/admin/adminProgram/chartAPI_stick.js
