+++ src/main/resources/templates/adminAsk/ask_write.html
@@ -6,9 +6,9 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" href="/css/common/font.css"> <!-- 폰트 -->
-    <link rel="stylesheet" href="/css/adminCommon/side_menu.css"> <!-- 사이드 메뉴 & 해더 -->
+    <link rel="stylesheet" href="/css/admin/adminCommon/side_menu.css"> <!-- 사이드 메뉴 & 해더 -->
     <!-- 글 작성 폼 -->
-    <link rel="stylesheet" href="/css/adminNotice/noticeWrite.css">
+    <link rel="stylesheet" href="/css/admin/adminNotice/noticeWrite.css">
 
     <!-- 썸머노트 API -->
     <link href="https://stackpath.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css" rel="stylesheet">
@@ -16,9 +16,9 @@
 
     <!-- 사이드 메뉴 -->
     <script src="http://code.jquery.com/jquery-latest.min.js" type="text/javascript"></script>
-    <script src="/js/adminCommon/side_menu.js"></script>
+    <script src="/js/admin/adminCommon/side_menu.js"></script>
     <!-- 버튼 클릭 시 확인 알림창 -->
-    <script src="/js/adminCommon/check.js"></script>
+    <script src="/js/admin/adminCommon/check.js"></script>
     <title>관리자-문의하기 답변 작성</title>
 </head>
 
@@ -26,7 +26,7 @@
 
 <div id="wrapper">
     <!-- 세로 메뉴바 -->
-    <nav th:replace="/fix/side_menu.html :: sideMenu"></nav>
+    <nav th:replace="fix/side_menu.html :: sideMenu"></nav>
     <!-- 헤더 ~ 밑에 메인 컨텐츠 -->
     <div id="container">
         <!-- 헤더 -->
@@ -95,5 +95,5 @@
 <!-- 썸머노트 -->
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/summernote@0.8.18/dist/summernote.min.js"></script>
-<script src="../js/summerNote.js"></script>
+<script src="/js/admin/summerNote.js"></script>
 <html>
