+++ src/main/resources/templates/adminMain/main.html
@@ -6,11 +6,11 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" href="/css/common/font.css"/>
-    <link rel="stylesheet" href="/css/adminCommon/side_menu.css"> <!-- 사이드 메뉴 & 해더 -->
-    <link rel="stylesheet" href="/css/adminMain/main_contents.css"> <!-- 메인 컨텐츠 -->
+    <link rel="stylesheet" href="/css/admin/adminCommon/side_menu.css"> <!-- 사이드 메뉴 & 해더 -->
+    <link rel="stylesheet" href="/css/admin/adminMain/main_contents.css"> <!-- 메인 컨텐츠 -->
     <!-- 사이드 메뉴 -->
     <script src="http://code.jquery.com/jquery-latest.min.js" type="text/javascript"></script>
-    <script src="/js/adminCommon/side_menu.js"></script>
+    <script src="/js/admin/adminCommon/side_menu.js"></script>
     <title>관리자-메인</title>
 </head>
 
@@ -346,5 +346,5 @@
 </body>
 <!-- 차트 -->
 <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
-<script src="/js/adminCommon/chartAPI.js"></script>
+<script src="/js/admin/adminCommon/chartAPI.js"></script>
 <html>
