+++ src/main/resources/templates/adminMentor/promotion.html
@@ -6,16 +6,16 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" href="/css/common/font.css"> <!-- 폰트 -->
-    <link rel="stylesheet" href="/css/adminCommon/side_menu.css"> <!-- 사이드 메뉴 & 해더 -->
-    <link rel="stylesheet" href="/css/adminCommon/contents.css"> <!-- 컨텐츠 공통 스타일 -->
-    <link rel="stylesheet" href="/css/adminCommon/page.css"> <!-- 페이징 -->
-    <link rel="stylesheet" href="/css/adminNotice/noticeForm.css"> <!-- 검색 폼 & 테이블 공통 스타일 -->
-    <link rel="stylesheet" href="/css/adminNotice/notice.css"> <!-- 테이블 개별 스타일 -->
+    <link rel="stylesheet" href="/css/admin/adminCommon/side_menu.css"> <!-- 사이드 메뉴 & 해더 -->
+    <link rel="stylesheet" href="/css/admin/adminCommon/contents.css"> <!-- 컨텐츠 공통 스타일 -->
+    <link rel="stylesheet" href="/css/admin/adminCommon/page.css"> <!-- 페이징 -->
+    <link rel="stylesheet" href="/css/admin/adminNotice/noticeForm.css"> <!-- 검색 폼 & 테이블 공통 스타일 -->
+    <link rel="stylesheet" href="/css/admin/adminNotice/notice.css"> <!-- 테이블 개별 스타일 -->
     <!-- 사이드 메뉴 -->
     <script src="http://code.jquery.com/jquery-latest.min.js" type="text/javascript"></script>
-    <script src="/js/adminCommon/side_menu.js"></script>
+    <script src="/js/admin/adminCommon/side_menu.js"></script>
     <!-- 버튼 클릭 시 확인 알림창 -->
-    <script src="/js/adminCommon/check.js"></script>
+    <script src="/js/admin/adminCommon/check.js"></script>
     <title>관리자-멘토홍보 글 관리</title>
 </head>
