+++ src/main/resources/templates/fix/side_menu.html
@@ -3,8 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
-    <link rel="stylesheet" href="/css/adminCommon/side_menu.css"> <!-- 사이드 메뉴 & 해더 -->
-    <script src="/js/adminCommon/side_menu.js"></script>
+    <link rel="stylesheet" href="/css/admin/adminCommon/side_menu.css"> <!-- 사이드 메뉴 & 해더 -->
+    <script src="/js/admin/adminCommon/side_menu.js"></script>
 </head>
 <body>
     <nav th:fragment="sideMenu">
